Remove skip_repeated_mv for NEW_INTER_MODES
skip_repeated_mv was primarily useful for handling the complex
interactions between NEARMV/NEARESTMV/GLOBALMV when it was possible for
all three modes to use the same MV.  Now, NEARESTMV is gone and it is no
longer trivial to check if NEARMV and GLOBALMV will use the same motion
vector.

Removing this optimization makes the total encoder complexity change for
NEW_INTER_MODES +5% with total BD-rate of -0.119% on 50-frame lowres.

Relative to NEW_INTER_MODES in the previous commit, the complexity change is +0.0075% and the BD-rate change is +0.02%.

This optimization very slightly _worsens_ BD-rate, but it's not correct with NEW_INTER_MODES and should be removed.

Change-Id: I5df6ba4cbf17ae2dcb232ec781a3009ca75b8c01