Enable warning C4800 for Visual Studio 2019.
Warning C4800 is a level 3 warning in Visual Studio 2015 and earlier.
Our Visual Studio nightly builds use Visual Studio 2015.

This warning is off by default starting in Visual Studio 2019. To debug
a Visual Studio nightly build failure caused by this warning, I wrote
this patch so I could reproduce the warning in Visual Studio 2019.

Note that casting an int expression to type bool doesn't disable the
warning, which is by design. This means this warning may be a nuisance
in some cases.

Change-Id: I63db86797f5e86a890a7f4125f46475e26992c84