Determine screen content type by encoding key frame twice
Summary:
-1% gain on screen content sets.

This cl makes better use of AV1 screen content (sc) tools.

To determine whether to use sc tools, this cl encodes the key frame
twice with a fixed partition and a high q, with and without sc tools.

The psnr is used to make decision:
If the psnr difference is significant, it indicates the video
is a screen content type, and we can safely use sc tools.

Speed and performance are evaluated:
(1). The recoding of the key frame adds little speed overhead, since
we use a fixed partition and a high q.

(2). We get -1% gain on screen content set. Because this cl
correctly classifies these videos as sc type than the old criteria.

STATS_CHANGED

Change-Id: I384c2c7b2670fa3b5fb128f67053775cde672133