Determine screen content type by encoding key frame twice
Summary:
-1.3% gain on screen content sets.
Add a speed feature which gains -0.1% to -0.4%. If turned on,
there's considerably large slowdown.

This cl makes better use of AV1 screen content (sc) tools.

To determine whether to use sc tools, this cl encodes the key frame
twice with a fixed partition and a high q, with and without sc tools.

The psnr and projected frame size is used to make decision:
(1). If the psnr difference is significant, it indicates the video
is a screen content type, and we can safely use sc tools.
(2). If the psnr difference is not significant, we may still use
sc tools, but guarded by a speed feature.

Speed and performance are evaluated:
(1). The recoding of the key frame adds little speed overhead, since
we use a fixed partition and a high q.

(2). We get -1.3% gain on screen content set. Because this cl
correctly classifies these videos as sc type than the old criteria.

(3). If we turn on the speed feature, more videos, including non sc
videos could be allowed to use sc tools. We can get -0.1% to -0.4%
gain. But it should be careful since the encoder will slow down up
to 50%, because of rd search for sc tools.  Conservatively, a speed
feature is added.  We only turn it on for speed 0 for now.

STATS_CHANGED

Change-Id: I384c2c7b2670fa3b5fb128f67053775cde672133