Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> (1 comment)
> 
> > Patch Set 1:
> > 
> > (1 comment)
> > 
> > The numbers are good.
> > 
> > Have you look at the perceptual quality at key frame? Note that unlike ARF frames, the key frame doesn't have overlay frame to make its display version aligned with the actual source (unfiltered) frame, at least for this moment.
> > 
> > Over filtering typically helps PSNR/SSIM. I am not sure about its impact on the perceptual quality front.
> > 
> > Paul, what do you think?
> 
> I have checked the visual quality this morning. To be honest, I cannot tell the difference between the original version and the new version with more filtering frames.

Responding to your comments Jingning and apologies for the length of this.

The issue with subjective visual quality vs metrics for this is complex and situation dependent.

With VP9 we ran into an interesting problem with high value, high resolution content. For example paid for movies streamed via Play Movies or YouTube Movies.

Here the data rates are quite a bit higher than for standard YouTube content and most really egregious visual artifacts are well controlled.

The exception to this is certain types of blocking and banding artifact especially in dark low contrast regions. These are invisible or barely visible on phones or normal computer monitors but when a video is played on a big screen HDR TV they can become very intrusive. However, they have very little impact on most metrics because the actual pixel errors are very small. In fact some metrics can report a perfect score when there are clear blocking or banding defects.

In VP9 we addressed this specific use case with a FILM mode setting that acts to try and preserve noise / film grain or the semblance of such. It does this by reducing the temporal filtering and taking a hit by spending more bits on key frames and ARF frames to actually code noise / grain. On other frames it then encourages the selection of modes (particularly inter modes) that retain a similar spatial complexity / texture to the source, even if these don’t yield the best psrn. From frame to frame your then get a reasonable simulation of the original noise at least in terms of look, feel and amplitude. This is very effective at masking the aforementioned blocking and banding artifacts and gives quite a dramatic improvement in subjective quality in some cases even whilst it hurts the traditional objective metrics.

For AV1 things get even more complicated. The bitstream supports a mechanism for modelling noise and film grain in the source, filtering it out prior to encoding and then using the model to reconstruct a similar pattern of noise in the final displayed video.  This has the potential to dramatically improve compression rates on noisy sources whilst retaining the visual benefits (especially on big screen TVs) of having some noise present in the final displayed video. From a metrics standpoint this may help PSNR (or other metrics) if  measured between the source (filtered or unfiltered) and the output frame prior to adding modeled noise, but the PSNR after adding back the pseudo noise will usually be much worse.

At the moment the use of this technique is still experimental but it may well negate the benefit from a metrics perspective of increasing temporal filtering as the whole source will be filtered. In cases where we cannot use this technique (for example if the decode hardware does not support the use of modeled noise) then we will likely have to revert for high value content to something similar to what we do in VP9. In that case we would not want to increase the key frame filtering as this will often remove the noise / grain that the process relies upon.

For lower bandwidths and resolutions  the subjective and objective measurements may be better aligned (VP9 FILM mode is not a good idea for such content) and this could well be worthwhile.

As such I think it is worth having this change available as an option and running some subjective quality tests. I don’t think it should yet be applied in all cases but even as things stand in this patch I think I am right in believing it will not apply to very high resolution and high data rate cases such as 4K HDR.

If and when we have a better understanding of the impact of pre-filtering and noise modeling it may be rendered redundant.

Paul

Patch-set: 1
Reviewer: Gerrit User 5610 <5610@e5514cf8-2d6e-3e29-adb4-24cd6dde4bf0>