Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> > Patch Set 2:
> > 
> > > Patch Set 2:
> > > 
> > > I don't think we should be wasting time on 5 year old compilers, but that's just me I guess.
> > 
> > I agree. Can we upgrade the Visual Studio nightly build to use Visual Studio 2019?
> 
> It already does 2015/2017/2019.

What I was trying to convey and failing miserably at is that I think we should drop 2015 from the tests. I can't think of a reason to justify continuing to spend cycles on triage/investigation/fixes of issues that occur only with 2015 when newer versions are freely available. TBH I don't think we need 2017 either, but at least that's only a few years old.

If we're really concerned about breaking an old MS compiler, and to be clear I don't think we should be (you could always upgrade to a newer platform SDK and compiler in the old IDEs, and I don't think that stopped), I think it would be more efficient to drive all the builds through one of the IDEs and use toolset configs to target the old compilers. This can be done using the -T arg or CMAKE_GENERATOR_TOOLSET[1]. I don't think we should do this, but if there are real concerns it would be a reasonable middle ground.

Still LGTM btw.

[1] https://cmake.org/cmake/help/latest/variable/CMAKE_GENERATOR_TOOLSET.html

Patch-set: 2
Reviewer: Gerrit User 5215 <5215@e5514cf8-2d6e-3e29-adb4-24cd6dde4bf0>