Regularize temporal filter's motion search with L1 norm
Previously, temporal filter used the rate cost of mv in the full pixel
search process, which can lead to some instability depending on how the
mv's cdf had adapted. This CL corrected the cost to L1 norm of the mv.

During full pixel motion search, the cost is now
  sad + \lambda * |mv|_1
where \lambda is a hard-coded value based on whether the video is lt
480p, geq 480p, or geq 720p. A better performance can be achieved by
adaptively picking the \lambda based on the frame size and motion
activity level.

Performance on Speed 1
 TESTSET | AVG_PSNR | OVR_PSNR |   SSIM  |
  LOWRES |  -0.049% |  -0.056% | -0.141% |
  MIDRES |  -0.169% |  -0.159% | -0.304% |
   HDRES |  -0.117% |  -0.118% | -0.183% |

STATS_CHANGED

Change-Id: I399fc0dc4836d3b16cdfe589802bb4c0b68302a9