Temporal filter: handle 10bit encoding and re-enable planewise strategy
The CL [1]. https://aomedia-review.googlesource.com/c/aom/+/102361
introduces a new function for planewise filtering and applies such
strategy on the key frame filtering.

But quality regression (2-3%) was reported on 10 bit input, and thus
reverted in this CL [2]:
https://aomedia-review.googlesource.com/c/aom/+/104188

This change scales down the patch difference calculated in non-local
means method, when the input is 10 bit.

It now brings back the gain of [1] on 8 bit input, as well as
obtaining gains on 10 bit input by xx%.

Performance on AWCY: (encode with "--bit-depth=10")

Change-Id: Ic1fc64a2a9308bfea8aa52049249157577a3a5b0