Temporal filter: handle 10bit encoding and re-enable planewise strategy
The CL [1]. https://aomedia-review.googlesource.com/c/aom/+/102361
introduces a new function for planewise filtering and applies such
strategy on the key frame filtering.

But two problems are reported:
(1). Quality regression (2-3%) on 10 bit input, and thus reverted
in this CL [2]:
https://aomedia-review.googlesource.com/c/aom/+/104188

(2). A huge quality regression (20%) of clip "dark720_60f" on 8 bit.

This change brings the planewise filtering for key frame back and
addresses the two problems:
(1). Scale down the patch difference calculated in non-local means
method, when the input is 10 bit.

(2). Slightly adjust the parameter to apply more smoothing towards
videos of very small noise (noise_level of this clip is 0.25).
The regression of the clip "dark720_60f" reduces from 20% to 2%,
Other videos remain mostly unchanged.

Results:
(1). This change now brings back the gain of [1] on 8 bit input.
(2). Instead of quality regression, it brings gains on 10 bit input:

Performance on AWCY: (encode with "--bit-depth=10")
   PSNR | PSNR Cb | PSNR Cr | PSNR HVS |    SSIM | MS SSIM | CIEDE 2000
-1.7614 | -2.1796 | -2.3363 |  -2.1234 | -2.1447 | -2.1953 |    -1.4159

STATS_CHANGED

Change-Id: Ic1fc64a2a9308bfea8aa52049249157577a3a5b0