Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> Mufaddal,
> Just wanted to confirm if this issue wouldn't happen in experimental?

Harish,
We have combined this fix with the commit which introduced the bug in experimental. We had to do this because experimental was showing a build fail without the fix combined into the commit. The test which failed was not run on master at the time we pushed this commit, hence it wasn't caught on master, so we had to push a fix later.

Patch-set: 2