Skip evaluating txfm rd of a mode based on skip rd
The speed feature txfm_rd_gate_level is introduced which gates the transform
rd evaluation of a mode if the skip rd of the mode is very high compared to
the best mode's skip rd. The speed feature is enabled for non-boosted frames
of non-screen contents in speed presets 4 and 5.

            Encode Time
cpu-used     Reduction      Quality Loss
   4           3.256%         0.1270%
   5           4.326%         0.2052%

STATS_CHANGED

Change-Id: I0669aac0255259871d8bf91140882bd424f38756