Add adaptive shrinking of gf group length.
Generally we want to place the altref on somewhere with lower
error/motion, so it could help more frames. This patch determines
the high/low errors regions from first pass stats, and uses it to
determine whether we want to shrink the gf group. Around 0.5% gain
on average PSNR and 0.4% gain on overall PSNR are observed on
speed4 test:

	avgPSNR	  overallPSNR 	 ssim		VMAF
lowres	-0.466	  -0.410	 -0.441		-0.487
midres	-0.335	  -0.325	 -0.347		-0.341
ugc360p	-0.746	  -0.551	 -0.455		-0.596

Speed1 test results are similar with a slightly smaller gain.

STATS_CHANGED

Change-Id: I89cdeba5483bd75b20ad090675fea1a715a07a0d