Add adaptive shrinking of gf group length.
Generally we want to place the altref on somewhere with lower
error/motion, so it could help more frames. This patch determines
the high/low errors regions from first pass stats, and uses it to
determine whether we want to shrink the gf group. Around 0.6% gain
on average PSNR and 0.4% gain on overall PSNR are observed on
speed4 test:

	avgPSNR	  overallPSNR 	 ssim		VMAF
lowres	-0.472	  -0.422	 -0.435		-0.454
midres	-0.649	  -0.351	 -0.276		 0.002
ugc360p	-0.788	  -0.566	 -0.492		-0.676

Speed1 test results are similar with a slightly smaller gain.

Change-Id: I89cdeba5483bd75b20ad090675fea1a715a07a0d