Update patch set 15
Patch Set 15:

I notice the use of MAX_GF_INTERVAL in defining various structures. Have you checked what happens if the user passes in a larger maximum GF interval on the command line. The #defined value is the maximum used by default in the existing pyramid structure but I am no so sure it is the maximum possible or that it is necessarily desirable for it to be so. Also does this still work and has it been tested for extended gf groups on static sections?

Patch-set: 15
Reviewer: Gerrit User 5610 <5610@e5514cf8-2d6e-3e29-adb4-24cd6dde4bf0>