Update patch set 9
Patch Set 9:

Hi Yunqing,

This patch introduces a speed feature 'txfm_rd_gate_level' to gate the transform rd evaluation of a mode if its skip rd turns out to be much higher than the best mode's skip rd. The aggressiveness of the gating is such that it is more aggressive for higher QPs and conservative for lower QPs or when the prediction is poor (i.e., the skip rd of the best mode is higher than the variance of the block). Since this approach prunes the less likely modes by gating the transform rd computation based on the skip rd, the speed feature is kept in 'inter_sf' and not 'tx_sf'.

This speed feature is enabled for non-boosted frames in non-screen contents for speed presets 4 and 5. The encode time reduction and BD rate impact in the commit message are obtained on the new test suite with 82 clips. It was verified locally that speed preset 3 was bit-exact with no impact on speed.

In the commit message, Quality Loss column:
 [+] sign is for BD-Rate drop
 [-] sign is for BD-Rate improvement
The instruction count reduction specified has been measured using command:
 $ perf stat -e instructions:u ./aomenc ...

Please review the patch.

Regards,
Nithya

Patch-set: 9
Reviewer: Gerrit User 5185 <5185@e5514cf8-2d6e-3e29-adb4-24cd6dde4bf0>
CC: Gerrit User 9039 <9039@e5514cf8-2d6e-3e29-adb4-24cd6dde4bf0>
CC-email: vikas.prasad@ittiam.com