Update patch set 15
Patch Set 15:

> Patch Set 15:
> 
> I notice the use of MAX_GF_INTERVAL in defining various structures. Have you checked what happens if the user passes in a larger maximum GF interval on the command line. The #defined value is the maximum used by default in the existing pyramid structure but I am no so sure it is the maximum possible or that it is necessarily desirable for it to be so. Also does this still work and has it been tested for extended gf groups on static sections?

Currently the when the gf group length is larger than active_max_gf_interval, we pretty much fall back to what was used before. The active_max_gf_interval is bounded by MAX_GF_INTERVAL so even if the user passes in a longer one we probably will still max out at 16. This may not be the optimal, though. 

With static sections, it is possible to exceed MAX_GF_INTERVALS, but then we will just fall back to the original mechanism. I don't particularly see any damage there but I also did not specifically test it. Could you let me know if there are any test sets or seqs that I can test it on?

Patch-set: 15