Update patch set 15
Patch Set 15:

The only other thing I noticed in testing this locally is that we often get odd group lengths like 13 which do not readily break down with the multi layer partition structure and end up with lopsided arrangements of ARF frames. For example one group with the following numbers of frames at each ARF levels 1, 1, 2, 1 vs a 16 frame group which would have 1, 1, 2, 4. I.e. there was a single odd level 4 ARF frame because of an odd number of frames. It might be worth looking at changes to either catch this and use extra leaf frames or restrict the size options for the group. It might also be nice to extend this work to consider groups larger than 16 frames if the stats point to this.

Overall though a good and useful gain.

Patch-set: 15
Reviewer: Gerrit User 5610 <5610@e5514cf8-2d6e-3e29-adb4-24cd6dde4bf0>