Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> > Patch Set 2:
> > 
> > Looks great.
> > 
> > Btw, did you figure out why the difference between 1-pass and 2-pass expanded on the master branch recently?
> 
> Another question, how much do the --kf-min-dist and --kf-max-dist parameters matter? What happens if we did not use them?

Hi Debargha,
Current fix is to address the BD rate gap between single pass and two pass on master branch (seen recently). 
In parent version, the gap is ~3.35%. With the fix, the gap is reduced to 0.25%. (Tests are done with kf-interval=60).

We see that kf-interval matters while combining tpl boost and lookahead based boost in the single pass. 
In single pass, if kf-interval is 1000 and number of frames encoded is 60, key/ARF frames are boosted more than expected (as projected boost would be based on 1000/144, but not 60). In two pass, as the total stats are known, key frame interval is taken as 60 and key/ARF frames are boosted appropriately. Hence, more quality gap between two pass and single pass is seen with kf-interval=1000 than kf-interval=60. 

As per our internal tests, quality gap between two pass and single pass = ~1.3% for kf-interval=1000; whereas the quality gap between two pass and single pass = ~0.25% for kf-interval=60 (for 60 frames limit)

We expect the quality gap to be ~1.3% (between single pass and two pass) due to the following reasons:
1. Default kf-max-dist is 9999
2. Boost combining logic clips the interval to 144 on high side. 

Regards,
Ranjit

Patch-set: 2