Fix decoder mismatches in the partition tool
- Force chroma plane only use the first reference frame for big(>=8x8)
blocks in a specially handled chroma reference group. The reason is
build_inter_predictors_sub8x8() operates in units of sub-4x4 blocks
and assumes all the blocks in the group uses a single ref mode.
If the conv_params are produced from the original prediction modes,
the decoder will produce nonsense predictions.
- Correct the check for block availability in ref mv scan

Change-Id: Icadfc941d7ceab169a3e259d5d6252864e3cc796