Update patch set 3
Patch Set 3:

> Patch Set 2:
> 
> Thanks Ranjit.
> 
> Can we cap the kf-max-dist as used for calculating boost to a more realistic number, say 150 or 300? Will that help narrow the gap ?

Debargha,
The quality gaps are due to mismatch between actual key frame interval (that can be shorter due to limit) and kf-max-dist. 
By setting default kf-max-dist to 150, the quality gap (between 2 pass and single pass) would be lesser if limit is 150 and would be more if limit is 60. 

Based on above observations, we feel following options are possible to reduce the quality gap.

1. oxcf->limit can be used to decide the actual key-frame-interval in single pass. As per our understanding, this approach would address majority of the quality gap and should help in bench-marking exercises and chunked encoding as limit is known beforehand. But oxcf->limit may not be populated correctly always (when encoder is integrated into other frameworks/--limit is not set). With this change, we expect quality gap to be closer to 0.2%

2. Modifying the boost weighing mechanism to be independent of key-frame-interval is another option. GFU boost weighing can be modified based on baseline_gf_interval in single pass. However kf_boost weighing will still be based on key-frame-interval and hence the quality gap would be bridged partially, but not entirely.

We are planning to look into approach 2 in above. 
Please let us know your opinion on our understanding.


Regards,
Ranjit

Patch-set: 3