Update patch set 10
Patch Set 10:

Debargha,
This patch fixes the weighing and calculation of gfu_boost for lower values of lag_in_frames in the case of LAP. As all first pass stats are not available, it is better to restrict the projection and combination of gfu_boost with tpl_boost to the current GOP, instead of using frames_to_key. 
For this, we are using the number of frames dependent on the current GF (forward + backward frames) to project the tpl boost. If the prior gfu_boost did not have enough first pass stats to calculate its boost, we are projecting that as well in the same fashion as tpl boost. The combination of the boosts is now happening on the number of stats used for calculating the prior boost, as it reflects the reliability of the prior boost. We have also made appropriate changes to the min and max bound values inside combine_prior_with_tpl_boost, to handle lower GOP lengths. We assumed that the previous 4.0 lower bound hardcoded in the combination function was the square root of the DEFAULT_GF_INTERVAL, hence we have changed that to the square root of current gf_interval.
With this change, quality difference for kf-1000, lag-in-frames=19, passes=1 vs. kf-1000, lag-in-frames=19, passes=2 went from -1.72% to -0.62% on the experimental branch.
Link: https://beta.arewecompressedyet.com/?job=Feb24-experimental-LAP19-TPLFactorFix3-2%402020-02-24T12%3A59%3A43.697Z&job=Feb21-experimental-2pass-lag25%402020-02-21T10%3A17%3A54.234Z
Please review.

Patch-set: 10
Reviewer: Gerrit User 5189 <5189@e5514cf8-2d6e-3e29-adb4-24cd6dde4bf0>
Reviewer: Gerrit User 5186 <5186@e5514cf8-2d6e-3e29-adb4-24cd6dde4bf0>
CC: Gerrit User 15030 <15030@e5514cf8-2d6e-3e29-adb4-24cd6dde4bf0>
CC: Gerrit User 6245 <6245@e5514cf8-2d6e-3e29-adb4-24cd6dde4bf0>
CC: Gerrit User 9039 <9039@e5514cf8-2d6e-3e29-adb4-24cd6dde4bf0>
Work-in-progress: false