rtc: Add source content metrics to nonrd_pickmode
Apply it to use_modeled_nonrd_cost feature for speed 8:
this reduces the avg. bdrate on low-resoln for that
feature from 2.9 to 1.7, and the max bdrate on the
mac_marco_moving clip reduced from 6 to 2.5.

Source variance/see only computed for sb_max = 64x64,
(vga and lower) since neon optimization is missing for
128x128.

Keep feature off for now.
Change-Id: I589cd59f375bd95f014fd6f9f6ebf15b3df74caa