Added alternate bit allocation method.
Added an alternate algorithm for distributing bits over ARFs.
This is based on a fixed distribution pattern and ratio of bits
allocated.

Also added stats to record the actual Q used at different
levels in the coding pyramid. In tests with strict rate control
this proved to be a good baseline for the lowest allowed Q
for the next level.

All fetaures disabled by default for now so there should be no
stats change for this patch.

Change-Id: I402773c5e5a36a8417f8b38f8ef536d37dc98ac0