Update patch set 5
Patch Set 5:

(3 comments)

> Patch Set 3:
> 
> Please give it a test in speed 1 for low/mid/ugc too. Thanks.

Results added. There's one seq in midres (soccer-4-cif) that performs badly for both speed. A close look suggests that the extended padding results (for this seq) in a different decision and somehow made it worse (esp. ssim) for two datapoints. 

But overall we still gain or at least remain neutral for gop of 32. For gop of 16 it should get better though.

Patch-set: 5