update EXT_QUANT to have two different options
update EXT_QUANT experiment to have two different options:
1: EXT_QUANT_72 extend the qp range to [0, 72] from [0,63] q_index to
qstep mapping for q_index in the range [0, 255] do not change. qstep
mapping is extrapolated to the [0, 287] range

2: EXT_QUANT_64 still keeps the qp in the range [0, 63] and q_index in
the range of [0, 255], but adjust scaling factor for q_index >= 224, i.e
last 32 q_index to qstep maping is adjusted to increase the qstep
values.

Change-Id: I301a472c35546306acfed7dc67de8158b5c46765