Reduce screen content psnr threshold slightly
On chinaspeed_xga, the PSNR difference between screen-content coded
keyframe and non-screen-content coded keyframe is about 1.0dB, which is
the current threshold. This causes some large BDRate fluctuations on the
video as more features are added.

Reducing this threshold to 0.9dB will help eliminating some of the
noise. No stats or speed change is observed on lowres, midres, hdres,
and screen-content set, but there stats_changed is possible.

BUG=aomedia:2600

STATS_CHANGED

Change-Id: I495c6602a2bd5b75104eb4ee7cc388acd0dcdfa2