rtc: Fix to skip mv mode in nonrd_pickmode
Existing condition was skipping mv mode (when it has
same motion vector as previous tested mode) but only
against NEARESTMV mode.

Remove this condition and use the skip_this_mv
check but without the restriction to 0 motion.
This will allow for NEWMV to be skipped if it has
same motion vector as NEARMV.

This doesn't happen often so speedup is small,
~1% for street180p at speed 8.

Neutral/negligible change in bdrate for speed >= 7.

Change-Id: If870e42f018362006fef65ea5e3c2bcdbe6e9697