Add preliminary support for RU parameter sharing.
Every time we evaluate wiener for an RU, we will add the RU's statistics to an accumulator.  We will evalutate the wiener filter computed with the current stats and full bitcost agains the shared filter and a single "is shared" bit.  Currently, this bit does not exist; it hopefully will soon.

If the shard filter is worse, we will create a new accumulator and push
the old one into an array.  Otherwise, we will continue as normal.

Then, once the first pass is done, all RUs have their Wiener
filters updated with the shared block statistics.

For more accurate costing, I have to see the way that the other patch will work.

Change-Id: I22d81f2f2edd89f6858cc2289b12ef7e896dd5e8