Add preliminary support for LR parameter sharing.
The goal of this patch is to add support for sharing Wiener filter loop
restoration parameters.  This will appear internally as several RUs sharing
identical Wiener parameters.

This process by which this is done is straightforward (and potentially
too simple):

- Maintain an accumulator for the Wiener covariance and cross-covariance
matrices.
- When processing a new block for the first time, compute its optimal
Wiener filter as normal.
- Then accumulate its covariance and cross-covariance matrices into the
accumulator and compute an aggregate Wiener filter.
- Compare the RD-cost of the optimal filter with an estimate of the
RD-cost of the aggregate filter.  The aggregate filter only requires a
single skip bit to be written to the bitstream; this is not included in
this patch and the cost of RESTORE_NONE is used instead.
- If the aggregate filter cost is higher than the cost of the optimal
filter, create new accumulators and push them into the vector of
accumulators.
- Either way, store a stats group id for each RU that can be used to
index into the accumulators.

Then, in a second pass, each RU goes to the accumulator that it has been
assigned to and recomputes its Wiener filter based on the statistics
inside.

This is not likely to be an optimal solution, and it may not work well at all because it does not evaluate many RUs together.

Change-Id: I22d81f2f2edd89f6858cc2289b12ef7e896dd5e8