Update patch set 3
Patch Set 3:

Tim sent the following comment via email, please take a look


aom_dsp/bitreadher.h:104

Do we need to keep this commented-out code? I know it was copied from
the original file, so if you want to defer cleaning it up to a separate
change, that is completely fine. I think if we do want to keep it, it
should be controlled by a named pre-processor #define instead of being
commented out (unlike now, where it is controlled by one in addition to
be commented out).

aom_dsp/bitreader.h:131

This test is not equivalent to the code in daala_read_symbol(). In that
code, if nsymbs != ref_nsymbs, then cdf_error was set to zero, and the
"cdf error" message would not be printed. It seems like these two should
be equivalent. That is, again, a pre-existing issue, so I am okay with
that not being addressed in this patch.

aom_dsp/bitreader.h:132

I do not believe that 32767 is correct here. Please see the
corresponding comment below in bitwriter.h.

aom_dsp/bitwriter.h:66

I do not believe that 32767 is correct here. The actual CDFs used by
od_ec_encode_cdf_q15() all end in 0, not 32767 (since they are actually
"inverse CDFs", as denoted by the icdf name for the parameter in
entenc.c). You can see this in the various AOM_CDF<n> macros defined in
prob.h, which all end in AOM_ICDF(CDF_PROB_TOP), i.e., 0 (plus one extra
0 to store the count of how often a CDF is used).

Patch-set: 3
Reviewer: Gerrit User 5005 <5005@e5514cf8-2d6e-3e29-adb4-24cd6dde4bf0>