Release hash-me table memory after recode loop
The memory for the hash table used to be deallocated in the next frame
to facilitate integer mv precision decision. The integer mv precision
decision was later changed to become independent of the hash table, and
the code for memory deallocation was remove with it. As a result, we
could leak memory during the encode.

Since the hash table is now only used by intrabc, this CL removes the
hash tables from the frame buffers, and instead only allocate one in
encode_frame_internal.

Change-Id: I9b8f3ee389680c93f6ec79c0fa1ae119964149d3