rtc-svc: Add set reference to SVC interface
Add set references to the SET_SVC_REF_FRAME_CONFIG
control, this allows user to completely specify
the layered pattern via the control, and avoids
the user of the flags AOM_EFLAG_NO_REF_.

Use this in the (3 spatial, 3 temporal) example
(mode >= 7) in the svc sample encoder.

Update unittests with this reference set.

Change-Id: I6b1403b1d00b2dd628714fb3161dbcba757ad416