Allow alt-ref when num of still frames is small
Previously when transition to still is detected, alt-ref is always not
allowed. The intention is that for a large group of still frames,
the key frame is already a perfect predictor.

However, there is a new case that only a small number of still frames
present in the video. Our current criteria can not differentiate
this case with the long still frames scenario. In this case the
previous keyframe or golden frame may not be a good predictor for
these small number of still images since it is before transition.

This change differentiates these two cases and allows to use an
alt-ref for a small number of still frames, and thus improves quality
of these frames.

The original intent is not changed. Alt-ref is still not used when
the keyframe can be a good predictor for these still images.

With the change, the gains (overall_psnr) are:
Animation_81ba: -12.28%
sc_SlideShow: -6.82%
sc_web_browsing: -4.38%
sc_map -3.12%

Only these four clips show different coding results than the baseline
for two pass encoding on 150 frames, speed 1, over lowres, midres,
hdres, ugc360p, and screen content sets.

STATS_CHANGED

Change-Id: Ia37c06b9e0b907f9d5e7036d3830d9351634fde0