Allow arf when num of still frames is small
Previously when transition to still is detected, arf is not used.
The number of still frames (hard coded as 5) is too small.
In this case the previous keyframe or golden frame may not be
a good predictor for these still images since it is before transition.

We'd like to not use arf when the keyframe can be a good predictor
for these still images. Therefore we use a new variable here
to indicate whether we can use arf for this keyframe group.

With the change, the gains (overall_psnr) are:
Animation_81ba: -12.28%
sc_SlideShow: -6.82%
sc_web_browsing: -4.38%
sc_map -3.12%

Only these four clips show different coding results than the baseline
for two pass encoding on 150 frames, speed 1, over lowres, midres,
hdres, ugc360p, and screen content sets.

Not that detect_transition_to_still() is still used to detect gf
group cut, there the hard coded value is still there and may be
changed later.

Change-Id: Ia37c06b9e0b907f9d5e7036d3830d9351634fde0