Replace 'cdef_preset' with bool 'cdef_transmitted'
The original array 'cdef_preset' in MACROBLOCKD was an integer array,
but was only used to check if it is initialized on not. However, it was
not trivial to understand that from the original code.

So, 'read_cdef' and 'write_cdef' were simplified and rewritten to make
this logic clear.

Also, added detailed documentation for the new array 'cdef_transmitted'.

BUG=aomedia:2610

Change-Id: I5dc3b1dd51b6520882498acff09b1003176b2c56