Update patch set 4
Patch Set 4:

Hi Yunqing,
We did a study to prune compound mode based on mode type. From our analysis we found that quality drop is minimal, if all extended compound modes were pruned. This experiment gave an instruction count reduction of 6.95% (for speed 5) with quality drop (not as per trade-off).

In order to improve the trade-off, neighbor block information is also considered before pruning. If reference frames of current block match with both left and above neighbors, the relevant extended compound mode is not pruned. With this change, the speed-up is 3.527% with a bd-rate drop of 0.1221% for speed 5.

We observed that speed feature is as per trade-off for speed 4 and hence extended the same for speed=4 as well. 
Once the patch is merged, we plan to extend the speed feature for speed 3 with less aggressiveness.

            Encode Time             BD-Rate Loss
cpu-used     Reduction     avg.psnr    ovr.psnr    ssim
    4         5.192%       0.1617%     0.1592%    0.1053%
    5         3.527%       0.1221%     0.1249%    0.0690%

 In the commit message, BD-Rate Loss column:
 [+] sign is for BD-Rate drop
 [-] sign is for BD-Rate improvement

 The instruction count reduction specified has been measured using command:
 $ perf stat -e instructions:u ./aomenc ...

Please review the patch. 
Thank you.

Patch-set: 4
Reviewer: Gerrit User 5185 <5185@e5514cf8-2d6e-3e29-adb4-24cd6dde4bf0>
CC: Gerrit User 6245 <6245@e5514cf8-2d6e-3e29-adb4-24cd6dde4bf0>
CC: Gerrit User 13265 <13265@e5514cf8-2d6e-3e29-adb4-24cd6dde4bf0>
CC: Gerrit User 9039 <9039@e5514cf8-2d6e-3e29-adb4-24cd6dde4bf0>
CC-email: vikas.prasad@ittiam.com